requirements.txt
python-dotenv
langchain
langgraph
openai
pandas
pydantic
duckdb
langchain-openai
pyarrow